img-73 Add support for RGB24:BANDSEQUENTIAL encoded images.
This change adds support for RGB24:BANDSEQUENTIAL encoded images by adding a new layer of abstraction to the rendering library.
It introduces the concept of an ImageModeHandler, which encapsulates the algorithm necessary for parsing the image data while
delegating the creation of individual pixels to an ImageRepresentationHandler. Future refactoring changes will retrofit existing
imaging rending code into this pattern.