[IMG-90] Convert over multiband support to the new render API.
The JITC samples provide RGB settings in IREPBAND, but this isn't actually required.
However this version only works if there are RGB bands.

Note that this commit partly addresses [IMG-74] in that we already have BANDSEQUENTIAL
mode support, and this adds the multi part for this. Verified in i_3301b.ntf test case