[IMG-122] Protect against over-long string and byte arrays.
The logic is that for "normal" strings, we will pad or truncate as needed.

For numbers, we'll zero pad, but if the number is too long to fit, we'll throw an
exception, because that would otherwise result in a (silently) corrupted file.

The exception also applies for strings and byte arrays that are supposed to be of
known length (because that will also result in corrupted file, and likely indicates
an implementation problem.