IMG-197 Parser fails completely when on NumberFormatException caused by bad TRE data.
This change modifies how the TREs are parsed so that an exception during parsing of any one TRE
won't cause parsing to stop. It relies on the TRE tag and length fields to be accurate. Given the
length of the TRE, it reads the appropriate number of bytes and creates a new NitfReader from
that. If parsing the TRE fails for any reason, then the outer NitfReader should be able to
continue normally, since the stream pointer will still be in the correct position.

This showed an error in two tests, which are also fixed as part of this commit.